requirements.txt
fastapi==0.115.11
langchain_ollama==0.3.0
motor==3.7.0
passlib==1.7.4
pydantic==2.10.6
PyJWT==2.10.1
PyJWT==2.10.1
